# src/main/resources/static/swagger-apis/openapi.yaml
...
openapi: 3.0.3
swagger: 2.0
info:
  title: Evently API
  description: This API manages events data for its React client Evently App.
  termsOfService: urn:tos
  contact: {}
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: base url
"host": http://localhost:8080
"basePath": /api/v1
...
paths:
...
  "/events":
    get:
      tags:
      - Event
      summary: Get all events
      operationId: getAllUsingGET
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
      - Event
      summary: Create an event
      operationId: createUsingPOST_1
      security:
      - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventReq"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/EventRes"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/events/{id}":
    get:
      tags:
      - Event
      summary: Get a single event by its id
      operationId: getEventByIdUsingGET
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
      - Event
      summary: Update an event
      operationId: updateUsingPUT
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventReqUpdate"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/EventRes"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
      - Event
      summary: Delete an event
      operationId: deleteUsingDELETE_1
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/EventRes"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/events?type={type}":
    get:
      tags:
      - Event
      summary: Get events by type
      operationId: getByTypeUsingGET
      security:
      - bearer: []
      parameters:
      - in: path
        name: type
        description: type
        required: true
        allowReserved: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/events?search={search}":
    get:
      tags:
      - Event
      summary: Get events by search
      operationId: getBySearchUsingGET
      security:
      - bearer: []
      parameters:
      - in: path
        name: search
        description: search
        required: true
        allowReserved: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/events?tag={tag}":
    get:
      tags:
      - Event
      summary: Get events by tag
      operationId: getByTagUsingGET
      security:
      - bearer: []
      parameters:
      - in: path
        name: tag
        description: tag
        required: true
        allowReserved: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found        
  "/joined-events":
    get:
      tags:
      - Event
      summary: Get authenticated user joined events
      operationId: getUserJoinedEventsUsingGET
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/published-events":
    get:
      tags:
      - Event
      summary: Get authenticated user published events
      operationId: getAuthPublishedEventsUsingGET
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/users/{id}/events":
    get:
      tags:
      - Event
      summary: Get user published events by its id
      operationId: getUserPublishedEventsUsingGET
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/EventRes"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/events/{id}/images":
    post:
      tags:
      - Image
      summary: Upload event image
      operationId: uploadUsingPOST
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ImageRes"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/images":
    get:
      tags:
      - Image
      summary: Get all images
      operationId: getAllUsingGET_2
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Image"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
      - Image
      summary: Delete image by its url
      operationId: deleteByUrlUsingDELETE
      security:
      - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ImageReqDelete"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/images/{id}":
    get:
      tags:
      - Image
      summary: Get a single image by its id
      operationId: getByIdUsingGET_1
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Image"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
      - Image
      summary: Delete image by its id
      operationId: deleteByIdUsingDELETE
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/auth-notifications":
    get:
      tags:
      - Notification
      summary: Get authenticated user notifications
      operationId: getAuthNotificatonsUsingGET
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Notification"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/auth-notifications/{id}":
    delete:
      tags:
      - Notification
      summary: Delete notification
      operationId: deleteUsingDELETE_2
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/auth-notifications/{id}/check":
    patch:
      tags:
      - Notification
      summary: Toggle check
      operationId: toggleCheckUsingPATCH
      security:
      - bearer: []
      parameters:
      - name: id
        in: path
        description: id
        required: true
        style: simple
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Message"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/notifications":
    get:
      tags:
      - Notification
      summary: Get all notifications
      operationId: getAllUsingGET_3
      security:
      - bearer: []
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Notification"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
components:
  securitySchemes:
    bearer:
      type: apiKey
      name: Authorization
      in: header
      description: Token de autenticación
  schemas:
...
    EventReq:
      title: EventReq
      required:
      - date
      - description
      - title
      - type
      type: object
      properties:
        date:
          type: string
          format: date-time
        description:
          maxLength: 2500
          minLength: 3
          type: string
        title:
          maxLength: 300
          minLength: 3
          type: string
        type:
          maxLength: 15
          minLength: 3
          type: string
    EventReqUpdate:
      title: EventReqUpdate
      required:
      - date
      - description
      - title
      - type
      type: object
      properties:
        date:
          type: string
          format: date-time
        description:
          maxLength: 2500
          minLength: 3
          type: string
        title:
          maxLength: 300
          minLength: 3
          type: string
        type:
          maxLength: 15
          minLength: 3
          type: string
    EventRes:
      title: EventRes
      type: object
      properties:
        date:
          type: string
        description:
          type: string
        hour:
          type: string
        id:
          type: integer
          format: int64
        images:
          type: array
          items:
            type: string
        location:
          type: string
        participant:
          type: boolean
        participantsCount:
          type: integer
          format: int32
        publisher:
          "$ref": "#/components/schemas/NestedUser"
        requirements:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        type:
          type: string
    Image:
      title: Image
      type: object
      properties:
        id:
          type: integer
          format: int64
        imgId:
          type: string
        imgUrl:
          type: string
        name:
          type: string
    ImageReqDelete:
      title: ImageReqDelete
      required:
      - url
      type: object
      properties:
        url:
          maxLength: 2147483647
          minLength: 7
          type: string
    ImageRes:
      title: ImageRes
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        url:
          type: string
...
    Message:
      title: Message
      type: object
      properties:
        message:
          type: string
...
    NestedUser:
      title: NestedUser
      type: object
      properties:
        avatar:
          type: string
        completeName:
          type: string
        username:
          type: string
    Notification:
      title: Notification
      type: object
      properties:
        checked:
          type: boolean
        createdAt:
          "$ref": "#/components/schemas/Timestamp"
        description:
          type: string
        id:
          type: integer
          format: int64
        subject:
          type: string
...
    Timestamp:
      title: Timestamp
      type: object
      properties:
        date:
          type: integer
          format: int32
        day:
          type: integer
          format: int32
        hours:
          type: integer
          format: int32
        minutes:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        nanos:
          type: integer
          format: int32
        seconds:
          type: integer
          format: int32
        time:
          type: integer
          format: int64
        timezoneOffset:
          type: integer
          format: int32
        year:
          type: integer
          format: int32
...